models/bilstm/requirements.txt
spacy
numpy
tensorflow-datasets
sklearn
